Initial Conversion of Roots from Bootstrap to Foundation
The initial port of foundation back to roots again.  This allows you to
use all of the markup and javascript of Foundation.

A few issues need to be worked out in the Roots Nav to make it work
with the Foundation markup, but initially it will at least display the
navigation items.